Remove old "drafts" of cmspdft.py
